1.3004: RapidApp 1.3004 + new RAPI_PSGI_FAST_EXIT feature
Still based on rapi/rapidapp:1.3002 but installs 1.3004 from CPAN as
well as latest Rapi::Blog 1.0101

Also added new env variable feature RAPI_PSGI_FAST_EXIT -- see README.md

+++ rapi_psgi_control.pl
@@ -136,7 +136,11 @@ while(1) {
     
       for my $sig (@exit_sigs) {
         $SIG{$sig} = sub {
-          print "\n\n  [caught SIG$sig -- shutting down]\n\n";
+          if($ENV{RAPI_PSGI_FAST_EXIT}) {
+            print "\n\n  [rapi/psgi: caught SIG$sig, fast exiting -- RAPI_PSGI_FAST_EXIT is set]\n\n";
+            exit;
+          }
+          print "\n\n  [rapi/psgi: caught SIG$sig -- shutting down]\n\n";
           $BgReq->DEMOLISH if ($BgReq);
           $BgReq = undef;
           kill $sig => $pid;
